Split Linux and nto code into separate mods
